+++ src/bytecode/jit/jitcache.rs
@@ -2,7 +2,7 @@ use std::{collections::HashMap, fmt, ops::{Deref, DerefMut}};
 
 use smallvec::SmallVec;
 
-use crate::{bytecode::bytecode::ValueMeta, gc::gc};
+use crate::gc::gc;
 
 /// specialised JIT = 引数タグ列ごとに別ポインタ
 #[derive(Clone, Debug, Default)]
@@ -27,11 +27,11 @@ impl JitCache {
 
 /// 自クレート固有の新型 ― 孤児規則を回避
 #[derive(Clone, PartialEq, Eq, Hash, Debug)]
-pub struct TagVec(SmallVec<[u8; 4]>);
+pub struct TagVec(SmallVec<[i64; 4]>);
 
 /// == 便利のために Deref を生やす ==
 impl Deref for TagVec {
-    type Target = SmallVec<[u8; 4]>;
+    type Target = SmallVec<[i64; 4]>;
     fn deref(&self) -> &Self::Target { &self.0 }
 }
 impl DerefMut for TagVec {
@@ -39,11 +39,11 @@ impl DerefMut for TagVec {
 }
 
 /// From 変換
-impl From<SmallVec<[u8; 4]>> for TagVec {
-    fn from(v: SmallVec<[u8; 4]>) -> Self { TagVec(v) }
+impl From<SmallVec<[i64; 4]>> for TagVec {
+    fn from(v: SmallVec<[i64; 4]>) -> Self { TagVec(v) }
 }
-impl From<&Vec<u8>> for TagVec {
-    fn from(v: &Vec<u8>) -> Self { TagVec(v.iter().copied().collect()) }
+impl From<&Vec<i64>> for TagVec {
+    fn from(v: &Vec<i64>) -> Self { TagVec(v.iter().copied().collect()) }
 }
 
 /// Display 実装は **この新型** に対して行う
@@ -58,8 +58,8 @@ impl fmt::Display for TagVec {
 }
 
 /* ───────── 1. FromIterator<u8> ───────── */
-impl FromIterator<u8> for TagVec {
-    fn from_iter<I: IntoIterator<Item = u8>>(iter: I) -> Self {
-        TagVec(SmallVec::<[u8; 4]>::from_iter(iter))
+impl FromIterator<i64> for TagVec {
+    fn from_iter<I: IntoIterator<Item = i64>>(iter: I) -> Self {
+        TagVec(SmallVec::<[i64; 4]>::from_iter(iter))
     }
 }
